Does a Batch User Creation From a CSV
Assumes you have a batch user account creation template in hand already and you are just pointing to it. This current solution is a robust skeleton from which to develop a real-world solution. The command New-ADUser is heavily under utilized and could use a sister command like Set-ADAccount to give the tool more depth of customization. This tool, like most of the functions included here, are diamonds in the rough. A real-world application and some polish will make this script shine.